api/requirements.txt
fastapi==0.128.5
uvicorn[standard]==0.40.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
